Update app.py
Erro durante a execução: While running get_bq_client(), a streamlit element is called on some layout block created outside the function. This is incompatible with replaying the cached effect of that element, because the referenced block might not exist when the replay happens.

How to fix this:

Move the creation of $THING inside get_bq_client().
Move the call to the streamlit element outside of get_bq_client().
Remove the @st.cache_resource decorator from get_bq_client().